requirements.txt

viam-sdk==0.45.2

typing-extensions
